+++ elm.json
@@ -1,7 +1,8 @@
 {
     "type": "application",
     "source-directories": [
-        "src"
+        "src/elm",
+        "src/elm/models"
     ],
     "elm-version": "0.19.1",
     "dependencies": {
