+++ public/index.html
@@ -42,6 +42,6 @@
         You need to enable JavaScript to run this app.
     </noscript>
     <div id="root"></div>
-    <script src="../src/index.js" type="module"></script>
+    <script src="../src/js/index.js" type="module"></script>
 </body>
 </html>
