+++ src/elm/models/LocalStorageValue.elm
