+++ src/elm/models/Note.elm
