+++ src/js/FileHandleManager.js
