+++ src/js/index.js
@@ -1,7 +1,7 @@
 import "regenerator-runtime";
-import "./styles/main.css";
-import "./styles/themes/dark.css";
-import "./styles/themes/white.css";
+import "../styles/main.css";
+import "../styles/themes/dark.css";
+import "../styles/themes/white.css";
 import registerServiceWorker from "./registerServiceWorker";
 import runElmApp from "./runElmApp";
